refactor: Rename `…events/api-gateway/lib/hack/disassociateUsagePlan.js`
