+++ lib/plugins/aws/package/compile/events/api-gateway/index.js
@@ -273,7 +273,7 @@ class AwsCompileApigEvents {
       'before:remove:remove': async () => {
         const globalValidate = require('../../../../lib/validate').validate;
         const disassociateUsagePlan =
-          require('./lib/hack/disassociateUsagePlan').disassociateUsagePlan;
+          require('./lib/hack/disassociate-usage-plan').disassociateUsagePlan;
 
         return BbPromise.bind(this).then(globalValidate).then(disassociateUsagePlan);
       },
