+++ test/unit/lib/plugins/aws/package/compile/events/api-gateway/index.test.js
@@ -8,7 +8,7 @@ const Serverless = require('../../../../../../../../../lib/serverless');
 const validate = require('../../../../../../../../../lib/plugins/aws/lib/validate');
 const getServiceState = require('../../../../../../../../../lib/plugins/aws/lib/get-service-state');
 const updateStage = require('../../../../../../../../../lib/plugins/aws/package/compile/events/api-gateway/lib/hack/updateStage');
-const disassociateUsagePlan = require('../../../../../../../../../lib/plugins/aws/package/compile/events/api-gateway/lib/hack/disassociateUsagePlan');
+const disassociateUsagePlan = require('../../../../../../../../../lib/plugins/aws/package/compile/events/api-gateway/lib/hack/disassociate-usage-plan');
 
 describe('AwsCompileApigEvents', () => {
   let awsCompileApigEvents;
